feat(product): 添加SPU积分信息保存功能
- 新增CouponFeginClient用于调用优惠券服务
- 在SpuInfoServiceImpl中集成积分信息保存逻辑
- 通过远程调用保存SPU的购买积分和成长积分
- 修改SpuSaveDto中的积分字段类型为BigDecimal
- 添加积分信息保存失败的异常处理机制

+++ gulimall-Parent/services/gulimall-product/src/main/java/com/xjz/gulimall/product/dto/SpuSaveDto.java
@@ -57,8 +57,8 @@ public class SpuSaveDto {
     }
     @Data
     public static class Bounds {
-        public int buyBounds;
-        public int growBounds;
+        public BigDecimal buyBounds;
+        public BigDecimal growBounds;
     }
     @Data
     public static class BaseAttrs {
